Fix float/int conversion problem. As reported here: https://github.com/syzygy1/Cfish/issues/53

+++ src/ntsearch.c
@@ -95,7 +95,7 @@ Value search_NonPV(Pos *pos, Stack *ss, Value alpha, Depth depth, int cutNode)
 #ifdef BIG_TT
   posKey = pos_key() ^ (Key)excludedMove;
 #else
-  posKey = pos_key() ^ (Key)(excludedMove << 16);
+  posKey = pos_key() ^ (Key)((int32_t)excludedMove << 16);
 #endif
   tte = tt_probe(posKey, &ttHit);
   ttValue = ttHit ? value_from_tt(tte_value(tte), ss->ply) : VALUE_NONE;
